pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vynth"
version = "1.0.0"
description = "Professional voice sampler synthesizer — record, edit, and play your voice via MIDI"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Vynth Team"}]
keywords = ["synthesizer", "sampler", "voice", "midi", "audio", "dsp"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Topic :: Multimedia :: Sound/Audio :: Sound Synthesis",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "PyQt6>=6.5.0",
    "pyqtgraph>=0.13.3",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.0",
    "python-rtmidi>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-qt>=4.2.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
vynth = "vynth.app:main"

[project.gui-scripts]
vynth-gui = "vynth.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
qt_api = "pyqt6"
addopts = "-v --tb=short"
